backend/requirements.txt
Flask
PyYAML
flask-cors
openai
jsonref